+++ app/models/sla_calendar_holiday_query.rb
@@ -75,7 +75,7 @@ class SlaCalendarHolidayQuery < Query
 
   # For Query Class
   def base_scope
-    self.queried_class.visible.where(statement)
+    self.queried_class.visible.with_references.where(statement)
   end
 
   def all_sla_calendar_values